requirements.txt
altgraph==0.17.4
getmac==0.9.4
macholib==1.16.3
netifaces==0.11.0
numpy==1.26.3
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.4
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pynmea2==1.19.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tk==0.1.0
tzdata==2023.4
